requirements.txt
gradio
llama-index
pymongo
openai
python-dotenv